+++ src/Calculators/Basic/OhmsLaw/OhmsLaw.cs
@@ -56,16 +56,14 @@ namespace NinjaCalc {
                         new NumberUnit("V", 1e0, NumberPreference.DEFAULT),
                         new NumberUnit("kV", 1e3),
                     },
-                    0.0,
-                    Direction_t.Input);
+                    Directions.Input,
+                    null);
 
             // Add validators
             this.Voltage.AddValidator(Validator.IsNumber(CalcValidationLevels.Error));
             this.Voltage.AddValidator(Validator.IsGreaterThanZero(CalcValidationLevels.Error));
 
-            this.CalcVars.Add(
-                "voltage",
-                this.Voltage);
+            this.CalcVars.Add(this.Voltage);
 
             //===============================================================================================//
             //============================================ CURRENT ==========================================//
@@ -90,15 +88,13 @@ namespace NinjaCalc {
                         new NumberUnit("mA", 1e-3),
                         new NumberUnit("A", 1e0, NumberPreference.DEFAULT),
                     },
-                    0.0,
-                    Direction_t.Input);
+                    Directions.Input,
+                    null);
 
             this.Current.AddValidator(Validator.IsNumber(CalcValidationLevels.Error));
             this.Current.AddValidator(Validator.IsGreaterThanZero(CalcValidationLevels.Error));
 
-            this.CalcVars.Add(
-                "current",
-                this.Current);
+            this.CalcVars.Add(this.Current);
 
             //===============================================================================================//
             //========================================== RESISTANCE =========================================//
@@ -123,21 +119,21 @@ namespace NinjaCalc {
                         new NumberUnit("MΩ", 1e6),
                         new NumberUnit("GΩ", 1e9),
                     },
-                    0.0,
-                    Direction_t.Output);
-
-            //this.Resistance.Direction = Direction_t.Output;
+                    Directions.Output,
+                    null);
+            
             this.Resistance.AddValidator(Validator.IsNumber(CalcValidationLevels.Error));
             this.Resistance.AddValidator(Validator.IsGreaterThanZero(CalcValidationLevels.Error));
 
-            this.CalcVars.Add(
-                "resistance",
-                this.Resistance);
+            this.CalcVars.Add(this.Resistance);
+
+            //===============================================================================================//
+            //============================================== FINAL ==========================================//
+            //===============================================================================================//
 
             this.FindDependenciesAndDependants();
-
-            // Bring the calculator into a default, stable state
             this.RecalculateAllOutputs();
+            this.ValidateAllVariables();
 
         }
     }