+++ src/Calculators/Pcb/TrackCurrentIpc2221A/TrackCurrentIpc2221ACalculator.cs
@@ -50,27 +50,27 @@ namespace NinjaCalc.Calculators.Pcb.TrackCurrentIpc2221A {
             //===============================================================================================//
             //========================================= TRACE CURRENT =======================================//
             //===============================================================================================//
-
-            //! @todo, Move these into the constructor for the base object?
             
            this.TraceCurrent = new CalcVarNumericalInput(
                 "traceCurrent",
                 view.TrackCurrentValue,
-                view.TrackCurrentUnits,                    
-                //this.CalcVars,                    
+                view.TrackCurrentUnits,                                                  
                 new NumberUnit[]{
                     new NumberUnit("mA", 1e-3),
                     new NumberUnit("A", 1e0, NumberPreference.DEFAULT),
                 },
-                0.0);
+                null);
 
-            // Add validators
+            //===== VALIDATORS =====//
             this.TraceCurrent.AddValidator(Validator.IsNumber(CalcValidationLevels.Error));
             this.TraceCurrent.AddValidator(Validator.IsGreaterThanZero(CalcValidationLevels.Error));
+            this.TraceCurrent.AddValidator(
+                new Validator(() => {
+                    return ((this.TraceCurrent.RawVal > 35.0) ? CalcValidationLevels.Warning : CalcValidationLevels.Ok);                                      
+                },
+                "Current is above recommended maximum (35A). Equation will not be as accurate (extrapolation will occur)."));
 
-            this.CalcVars.Add(
-                "traceCurrent",
-                this.TraceCurrent);
+            this.CalcVars.Add(this.TraceCurrent);
 
             //===============================================================================================//
             //========================================== TEMP RISE ==========================================//
@@ -79,20 +79,27 @@ namespace NinjaCalc.Calculators.Pcb.TrackCurrentIpc2221A {
             this.TempRise = new CalcVarNumericalInput(
                 "tempRise",
                 view.TempRiseValue,
-                view.TempRiseUnits,                    
-                //this.CalcVars,                    
+                view.TempRiseUnits,                                                
                 new NumberUnit[]{
                     new NumberUnit("C", 1e0, NumberPreference.DEFAULT),                        
                 },
-                0.0);
+                null);
 
-            // Add validators
+            //===== VALIDATORS =====//
             this.TempRise.AddValidator(Validator.IsNumber(CalcValidationLevels.Error));
             this.TempRise.AddValidator(Validator.IsGreaterThanZero(CalcValidationLevels.Error));
+            this.TempRise.AddValidator(
+                new Validator(() => {
+                    return ((this.TempRise.RawVal < 10.0) ? CalcValidationLevels.Warning : CalcValidationLevels.Ok);
+                },
+                "Temperature rise is below the recommended minimum (10°C). Equation will not be as accurate (extrapolation will occur)."));
+            this.TempRise.AddValidator(
+                new Validator(() => {
+                    return ((this.TempRise.RawVal > 100.0) ? CalcValidationLevels.Warning : CalcValidationLevels.Ok);
+                },
+                "Temperature rise is above the recommended maximum (100°C). Equation will not be as accurate (extrapolation will occur)."));
 
-            this.CalcVars.Add(
-                "tempRise",
-                this.TempRise);
+            this.CalcVars.Add(this.TempRise);
 
             //===============================================================================================//
             //======================================== TRACK THICKNESS ======================================//
@@ -101,27 +108,33 @@ namespace NinjaCalc.Calculators.Pcb.TrackCurrentIpc2221A {
             this.TrackThickness = new CalcVarNumericalInput(
                 "trackThickness",
                 view.TrackThicknessValue,
-                view.TrackThicknessUnits,
-                //this.CalcVars,
+                view.TrackThicknessUnits,                
                 new NumberUnit[]{
                     new NumberUnit("um", 1e-6, NumberPreference.DEFAULT),                        
                     new NumberUnit("mm", 1e-3),                        
                 },
-                0.0);
+                null);
 
-            // Add validators
+            //===== VALIDATORS =====//
             this.TrackThickness.AddValidator(Validator.IsNumber(CalcValidationLevels.Error));
             this.TrackThickness.AddValidator(Validator.IsGreaterThanZero(CalcValidationLevels.Error));
+            this.TrackThickness.AddValidator(
+               new Validator(() => {
+                   return ((this.TrackThickness.RawVal < 17.5e-6) ? CalcValidationLevels.Warning : CalcValidationLevels.Ok);
+               },
+               "Track thickness is below the recommended minimum (17.5um or 0.5oz). Equation will not be as accurate (extrapolation will occur)."));
+            this.TrackThickness.AddValidator(
+                new Validator(() => {
+                    return ((this.TrackThickness.RawVal > 105.0036e-6) ? CalcValidationLevels.Warning : CalcValidationLevels.Ok);
+                },
+                "Track thickness is above the recommended maximum (105um or 3oz). Equation will not be as accurate (extrapolation will occur)."));
 
-            this.CalcVars.Add(
-                "trackThickness",
-                this.TrackThickness);
+            this.CalcVars.Add(this.TrackThickness);
 
             //===============================================================================================//
             //========================================= TRACK LAYER =========================================//
             //===============================================================================================//
 
-            // This is a combobox!!! How do we do this?
             this.TrackLayer = new CalcVarComboBox(
                 "trackLayer",
                 view.TrackLayer,
@@ -130,9 +143,7 @@ namespace NinjaCalc.Calculators.Pcb.TrackCurrentIpc2221A {
                     "External",
                 });
 
-            this.CalcVars.Add(
-                "trackLayer",
-                this.TrackLayer);
+            this.CalcVars.Add(this.TrackLayer);
 
             //===============================================================================================//
             //======================================== MIN. TRACK WIDTH =====================================//
@@ -142,15 +153,13 @@ namespace NinjaCalc.Calculators.Pcb.TrackCurrentIpc2221A {
                 "minTrackWidth",
                 view.MinTrackWidthValue,
                 view.MinTrackWidthUnits,
-                //this.CalcVars,
                 () => {
-                    Console.WriteLine("Equation() called for MinTrackWidth.");
+                    //Console.WriteLine("Equation() called for MinTrackWidth.");
                     var traceCurrent = this.TraceCurrent.RawVal;
                     var tempRise = this.TempRise.RawVal;
                     var trackThickness = this.TrackThickness.RawVal;
                     var trackLayer = this.TrackLayer.RawVal;
-
-                    //var trackLayer = TrackLayer_t.INTERNAL or TrackLayer_t.EXTERNAL
+                    
                     if(trackLayer == "External")     
 			        {
 				        Console.WriteLine("External trace selected.");
@@ -166,10 +175,11 @@ namespace NinjaCalc.Calculators.Pcb.TrackCurrentIpc2221A {
                         Console.WriteLine("Cross-sectional area = " + crossSectionalArea.ToString());
 				        double width = (crossSectionalArea/(trackThickness*1000000.0/25.4))*(25.4/1000000.0);
 				        return width;
-			        }
- 
-                    // EQUATION GOES HERE
-                    return 0.0;
+                    }
+                    else {
+                        System.Diagnostics.Debug.Assert(false, "Track layer was invalid (should be either External or Internal).");
+                        return Double.NaN;
+                    }
                 },
                 new NumberUnit[]{
                     new NumberUnit("um", 1e-6),                        
@@ -180,12 +190,30 @@ namespace NinjaCalc.Calculators.Pcb.TrackCurrentIpc2221A {
             this.MinTrackWidth.AddValidator(Validator.IsNumber(CalcValidationLevels.Error));
             this.MinTrackWidth.AddValidator(Validator.IsGreaterThanZero(CalcValidationLevels.Error));
 
-            this.CalcVars.Add(
-                "minTrackWidth",
-                this.MinTrackWidth);
+            this.CalcVars.Add(this.MinTrackWidth);
+
+            //===============================================================================================//
+            //=========================================== VIEW CONFIG =======================================//
+            //===============================================================================================//
+
+            // Setup the top PCB layer to dissappear if "External" is selected for the track layer,
+            // and visible if "Internal" is selected.
+            this.TrackLayer.RawValueChanged += (sender, e) => {
+                if (this.TrackLayer.RawVal == "Internal") {
+                    view.TopPcb.Visibility = System.Windows.Visibility.Visible;
+                }
+                else if (this.TrackLayer.RawVal == "External") {
+                    view.TopPcb.Visibility = System.Windows.Visibility.Collapsed;
+                }
+            };
+
+            //===============================================================================================//
+            //============================================== FINAL ==========================================//
+            //===============================================================================================//
 
             this.FindDependenciesAndDependants();
             this.RecalculateAllOutputs();
+            this.ValidateAllVariables();
             
         }       
     }