+++ src/Core/CalcVarBase.cs
@@ -25,6 +25,19 @@ namespace NinjaCalc.Core {
         }
 
         /// <summary>
+        /// Use this to subscribe to the calculator variables "raw" value changing.
+        /// </summary>
+        public event EventHandler RawValueChanged;
+
+        protected virtual void OnRawValueChanged(EventArgs e) {
+            EventHandler handler = RawValueChanged;
+            if (handler != null) {
+                handler(this, e);
+            }
+        }
+
+
+        /// <summary>
         /// Designed to be assigned to when Calculator.CalculateDependencies() is run. This is not calculated in this class's constructor,
         /// but rather once all calculator variables and their equations have been added to the calculator.
         /// </summary>
@@ -60,7 +73,7 @@ namespace NinjaCalc.Core {
             set;
         }
 
-        public virtual Direction_t Direction {
+        public virtual Directions Direction {
             get;
             set;
         }
@@ -88,7 +101,7 @@ namespace NinjaCalc.Core {
             Console.WriteLine("ForceDependantOutputsToRecalculate() called.");
             // We need to re-calculate any this calculator variables dependants, if they are outputs
             for (int i = 0; i < this.Dependants.Count; i++) {
-                if (this.Dependants[i].Direction == Direction_t.Output) {
+                if (this.Dependants[i].Direction == Directions.Output) {
                     this.Dependants[i].Calculate();
                 }
             }